data-pipeline/app/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.2
streamlit==1.30.0
pandas==2.2.0
requests==2.31.0
plotly==5.18.0
mysql-connector-python==8.0.33
cassandra-driver==3.29.0  # <--- THÊM DÒNG NÀY